Revert "LPS-147480 Creation of BaseNotificationsInfo"
This reverts commit c686d0e24fad6193ae5f9e1a2e899be0115a9781.
